:new: Playlists library! (#518)
Work in progress
 * Tidy Libraries up into more modules, was way too big
 * Add a playlist library which listens (a bit) to the main songs file library
 * link via `.playlists` property on main library, like with `.albums` (a bit)
 * Make playlists browser have fewer class methods and class singletony things
 * Rewire / rewrite various bits of Playlists browser to be less coupled and more listening to the new library for changes.
 * This solves #3283 and a few other gotchas, and actually simplifies logic it seems
 * Add a bit more (Generic) typing for collections and stuff
